% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSHEEP.R
\name{bring_grass}
\alias{bring_grass}
\title{get_block}
\usage{
bring_grass(
  sheep = NULL,
  id = "",
  height = 0,
  width = 0,
  label = "",
  overwrite_by_id = FALSE,
  plan = NULL,
  verbose = FALSE
)
}
\arguments{
\item{SHEEP_group}{...}
}
\value{
...
}
\description{
...
}
\examples{
...
}
