% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_color.R
\name{compute_color}
\alias{compute_color}
\title{Compute color}
\usage{
compute_color(
  value,
  min,
  max,
  Palette,
  colorStep = 256,
  include = FALSE,
  reverse = FALSE
)
}
\description{
Compute color
}
