% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{compute_colorBin}
\alias{compute_colorBin}
\title{Compute color bin}
\usage{
compute_colorBin(
  min,
  max,
  colorStep,
  center = NULL,
  include = FALSE,
  round = TRUE
)
}
\description{
Compute color bin
}
