% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_datasheet.R
\name{datasheet_ASHES}
\alias{datasheet_ASHES}
\title{Datasheet panel}
\usage{
datasheet_ASHES(
  list_df2plot,
  meta,
  trend_period,
  mean_period,
  linetype_per,
  axis_xlim,
  colorForce,
  exXprob,
  info_header,
  time_header,
  foot_note,
  structure,
  info_height,
  time_height,
  var_ratio,
  foot_height,
  paper_size,
  shapefile_list,
  logo_path,
  zone_to_show,
  show_colorEvent,
  outdirTmp_pdf,
  outdirTmp_png,
  df_page = NULL,
  pdf_chunk = "all"
)
}
\description{
Datasheet panel
}
