% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_color.R
\name{get_color}
\alias{get_color}
\title{Get color}
\usage{
get_color(
  value,
  min,
  max,
  Palette,
  colorStep = 256,
  reverse = FALSE,
  include = FALSE,
  noneColor = "black"
)
}
\description{
Get color
}
