% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSHEEP.R
\name{get_power}
\alias{get_power}
\title{Number formatting}
\usage{
get_power(value)
}
\arguments{
\item{text}{...}

\item{nLim}{...}

\item{20}{...}

\item{newlineId}{... (default : "\n")}
}
\value{
...
}
\description{
...
}
\examples{
...
}
