% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSHEEP.R
\name{gpct}
\alias{gpct}
\title{Pourcentage of variable}
\usage{
gpct(pct, L, min_lim = NULL, shift = FALSE)
}
\arguments{
\item{fig}{...}

\item{path}{...}
}
\value{
...
}
\description{
...
}
\examples{
...
}
