% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSHEEP.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{find_id}
\usage{
is.wholenumber(X, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{X}{...}

\item{a}{...}

\item{where}{... (default : "")}
}
\value{
...
}
\description{
...
}
\examples{
...
}
