% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_doc.R
\name{layout_panel}
\alias{layout_panel}
\title{Layout panel}
\usage{
layout_panel(
  data,
  meta,
  PLOT,
  structure,
  to_plot = c("datasheet_ASHES", "table_ASHES", "map_ASHES", "map_regime_ASHES",
    "map_trend_ASHES", "map_mean_ASHES"),
  tmpdir = "",
  figdir = "",
  docdir = "doc",
  level = 0.1,
  trend_period = NULL,
  mean_period = NULL,
  colorForce = FALSE,
  exXprob = 0.01,
  linetype_per = "solid",
  axis_xlim = NULL,
  paper_size = "A4",
  time_header = NULL,
  info_header = NULL,
  foot_note = TRUE,
  info_height = 2.8,
  time_height = 3,
  var_ratio = 3,
  foot_height = 1.25,
  shapefile_list = NULL,
  resdir = NULL,
  logo_path = NULL,
  zone_to_show = NULL,
  pdf_chunk = c("all"),
  show_colorEvent = FALSE
)
}
\description{
Layout panel
}
