% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_tools.R
\name{merge_panel}
\alias{merge_panel}
\title{Merge}
\usage{
merge_panel(
  STOCK,
  direction = "V",
  NAME = NULL,
  page_margin = c(t = 0, r = 0, b = 0, l = 0),
  paper_size = c(10, 10),
  hjust = 0,
  vjust = 1
)
}
\description{
Merge
}
