% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_info.R
\name{panel_info}
\alias{panel_info}
\title{Info panel}
\usage{
panel_info(
  list_df2plot,
  meta,
  trend_period = NULL,
  mean_period = NULL,
  period = NULL,
  shapefile_list = NULL,
  codeLight = NULL,
  data_code = NULL,
  to_do = "all",
  zone_to_show = "France"
)
}
\description{
Info panel
}
