% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_spaghetti.R
\name{panel_spaghetti}
\alias{panel_spaghetti}
\title{panel spaghetti}
\usage{
panel_spaghetti(
  data_code,
  Colors = NULL,
  title = "débit journalier",
  unit = "m^{3}.s^{-1}",
  alpha = 0.7,
  isSqrt = FALSE,
  missRect = FALSE,
  isBack = TRUE,
  isTitle = TRUE,
  dTitle = 0,
  sizeYticks = 9,
  date_labels = "\%Y",
  breaks = "10 years",
  minor_breaks = "2 years",
  d_breaks = 0,
  break_round = -1,
  isBackObsAbove = TRUE,
  axis_xlim = NULL,
  grid = TRUE,
  margin_add = margin(t = 0, r = 0, b = 0, l = 0, "mm"),
  first = FALSE,
  last = FALSE
)
}
\description{
panel spaghetti
}
