% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_trend.R
\name{panel_trend}
\alias{panel_trend}
\title{Time panel}
\usage{
panel_trend(
  dataEx_code,
  trend_code,
  var,
  unit,
  samplePeriod_code = NULL,
  linetype_per = "solid",
  level = 0.1,
  colorForce = FALSE,
  missRect = FALSE,
  unit2day = 365.25,
  trend_period = NULL,
  mean_period = NULL,
  axis_xlim = NULL,
  grid = TRUE,
  ymin_lim = NULL,
  color = NULL,
  first = FALSE,
  last = FALSE,
  lim_pct = 10
)
}
\description{
Time panel
}
