% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSHEEP.R
\name{plan_of_flock}
\alias{plan_of_flock}
\title{add_sheep}
\usage{
plan_of_flock(flock, plan)
}
\arguments{
\item{flock}{...}

\item{sheep}{...}

\item{NULL}{...}

\item{id}{... (default : "")}

\item{height}{...}

\item{width}{...}

\item{0}{...}

\item{label}{... (default : "")}

\item{overwrite_by_id}{... (default : FALSE)}
}
\value{
...
}
\description{
...
}
\examples{
...
}
