% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSHEEP.R
\name{return_to_sheepfold}
\alias{return_to_sheepfold}
\title{Merge}
\usage{
return_to_sheepfold(
  flock,
  page_margin = c(t = 0, r = 0, b = 0, l = 0),
  paper_size = NULL,
  hjust = 0,
  vjust = 1,
  verbose = FALSE
)
}
\arguments{
\item{flock}{...}

\item{sheep}{...}

\item{NULL}{...}

\item{id}{... (default : "")}

\item{0}{...}

\item{label}{... (default : "")}

\item{overwrite_by_id}{...}

\item{FALSE}{...}

\item{plan}{...}

\item{height}{...}

\item{TRUE}{...}

\item{width}{...}
}
\value{
...

...

...
}
\description{
...

...

...
}
\examples{
...
...
...
}
