% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSHEEP.R
\name{round_label}
\alias{round_label}
\title{round_label}
\usage{
round_label(labelRaw, direction = "V", ncharLim = 4)
}
\arguments{
\item{alphabet}{... (default : c()}

\item{letters}{...}

\item{LETTERS}{...}

\item{other_letters}{...}

\item{numbers}{... (default : , symbols))}

\item{size}{...}

\item{50}{...}

\item{font}{... (default : "sans")}

\item{style}{... (default : "normal")}

\item{isNorm}{...}

\item{TRUE}{...}

\item{out_dir}{... (default : "letter")}

\item{save}{... (default : FALSE)}
}
\value{
...
}
\description{
...
}
\examples{
...
}
