% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{theme_IPCC}
\alias{theme_IPCC}
\title{Ggplot2 theme ash}
\usage{
theme_IPCC(
  is_panel.background = FALSE,
  is_plot.title = TRUE,
  plot.title_size = 10,
  isGridX = FALSE,
  isGridY = TRUE,
  is_axis.line.x = TRUE,
  is_axis.ticks.x = TRUE,
  is_axis.text.x = TRUE,
  axis.text.x_size = 9,
  axis.text.x_angle = NULL,
  axis.text.x_hjust = NULL,
  axis.text.x_vjust = NULL,
  is_axis.ticks.y = TRUE,
  is_axis.text.y = TRUE,
  axis.text.y_margin = NULL,
  axis.text.y_size = 8,
  axis.text.y_vjust = 0.59,
  axis.ticks.length.y = 1.5,
  isLabelX = FALSE,
  isLabelY = FALSE,
  is_border = FALSE,
  family = "DejaVu Sans"
)
}
\description{
Ggplot2 theme ash
}
