% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSHEEP.R
\name{add_sheep}
\alias{add_sheep}
\title{load_font}
\usage{
add_sheep(
  flock,
  sheep = NULL,
  id = "",
  height = 0,
  width = 0,
  label = "",
  overwrite_by_id = FALSE
)
}
\arguments{
\item{path}{...}

\item{NULL}{...}

\item{force_import}{... (default : FALSE)}
}
\value{
...
}
\description{
...
}
\examples{
...
}
