% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSHEEP.R
\name{shear_sheeps}
\alias{shear_sheeps}
\title{is.wholenumber}
\usage{
shear_sheeps(flock, height = TRUE, width = TRUE)
}
\arguments{
\item{tol}{... (default : .Machine$double.eps^0.5)}

\item{x}{...}

\item{X}{...}

\item{den}{...}
}
\value{
...

...

...
}
\description{
...

...

...
}
\examples{
...
...
...
}
